Refactor MaintenanceBanner into folder with external styles
Move the component under Navbar/MaintenanceBanner/, extract Material-UI
styles to styles.js, use arrow functions for the component and
mapStateToProps, and update imports in Navbar and tests.
