+++ src/react/src/components/Navbar/Navbar.jsx
@@ -6,7 +6,7 @@ import { MobileNavbarItems, NavbarItems } from '../../util/constants';
 import Logo from './Logo';
 import BurgerButton from './BurgerButton';
 import NavbarQ42022Banner from './NavbarQ42022Banner';
-import MaintenanceBanner from './MaintenanceBanner';
+import MaintenanceBanner from './MaintenanceBanner/MaintenanceBanner';
 import MenuClickHandlerContext from './MenuClickHandlerContext';
 
 const breakpoint = '(max-width: 75rem)';
